refactor(vscode): disable noUnusedLocals and update build script. Extract InlineEditController
