+++ clients/vscode/src/extension.ts
@@ -88,9 +88,9 @@ export async function activate(context: ExtensionContext) {
 
   const issues = new Issues(client, config);
   const contextVariables = new ContextVariables(client, config);
-  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */ /* @ts-expect-error noUnusedLocals */
+  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */ // @ts-ignore noUnusedLocals
   const statusBarItem = new StatusBarItem(context, client, config, issues, inlineCompletionProvider);
-  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */ /* @ts-expect-error noUnusedLocals */
+  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */ // @ts-ignore noUnusedLocals
   const commands = new Commands(
     context,
     client,
